* Moved several functions from ApiBlock to ApiWorld, since they're working with the world not just blocks + Added an option to ToolRecipes to change the material used for sticks + Added NONE to UpdateType
