+++ src/com/blazeloader/api/toolset/ToolRecipes.java
@@ -1,7 +1,12 @@
 package com.blazeloader.api.toolset;
 
+import java.util.List;
+
 import com.blazeloader.api.recipe.ApiCrafting;
 import com.blazeloader.api.recipe.ApiCrafting.BLCraftingManager;
+import com.blazeloader.bl.obf.BLOBF;
+import com.google.common.collect.Lists;
+
 import net.minecraft.init.Items;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
@@ -9,21 +14,28 @@ import net.minecraft.item.crafting.RecipesTools;
 
 public class ToolRecipes {
     private static final String[][] patterns = new RecipesTools().recipePatterns;
-
+    
     private final BLCraftingManager manager;
-
+	
+    private Item stickMaterial = Items.stick;
+    
     public ToolRecipes() {
         this(ApiCrafting.getVanillaCraftingManager());
     }
 
     public ToolRecipes(BLCraftingManager CraftingManager) {
         manager = CraftingManager;
+        AddToolSetRecipes(Items.stick, Items.stick);
     }
 
-    public void AddToolSetRecipes(Item material, Item... tools) {
+    public void AddToolSetRecipes(Item toolMaterial, Item... tools) {
         for (int i = 0; i < tools.length && i < patterns.length; i++) {
-            manager.addRecipe(new ItemStack(tools[i]),
-                    patterns[i], '#', Items.stick, 'X', material);
+            manager.addRecipe(new ItemStack(tools[i]), patterns[i], '#', stickMaterial, 'X', toolMaterial);
         }
+    }
+    
+    public ToolRecipes setStick(Item stick) {
+    	stickMaterial = stick;
+    	return this;
     }
 }
